chg: Refactored tests
